lambda/get-prompt-suggestions/requirements.txt
openai
pydantic-core